push utils.Copy(File|Dir) -> model
