+++ server/channels/utils/file_test.go
@@ -7,63 +7,11 @@ import (
 	"bytes"
 	"crypto/rand"
 	"io"
-	"os"
-	"path/filepath"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func TestCopyDir(t *testing.T) {
-	srcDir, err := os.MkdirTemp("", "src")
-	require.NoError(t, err)
-	defer os.RemoveAll(srcDir)
-
-	dstParentDir, err := os.MkdirTemp("", "dstparent")
-	require.NoError(t, err)
-	defer os.RemoveAll(dstParentDir)
-
-	dstDir := filepath.Join(dstParentDir, "dst")
-
-	tempFile := "temp.txt"
-	err = os.WriteFile(filepath.Join(srcDir, tempFile), []byte("test file"), 0655)
-	require.NoError(t, err)
-
-	childDir := "child"
-	err = os.Mkdir(filepath.Join(srcDir, childDir), 0777)
-	require.NoError(t, err)
-
-	childTempFile := "childtemp.txt"
-	err = os.WriteFile(filepath.Join(srcDir, childDir, childTempFile), []byte("test file"), 0755)
-	require.NoError(t, err)
-
-	err = CopyDir(srcDir, dstDir)
-	assert.NoError(t, err)
-
-	stat, err := os.Stat(filepath.Join(dstDir, tempFile))
-	assert.NoError(t, err)
-	assert.Equal(t, uint32(0655), uint32(stat.Mode()))
-	assert.False(t, stat.IsDir())
-	data, err := os.ReadFile(filepath.Join(dstDir, tempFile))
-	assert.NoError(t, err)
-	assert.Equal(t, "test file", string(data))
-
-	stat, err = os.Stat(filepath.Join(dstDir, childDir))
-	assert.NoError(t, err)
-	assert.True(t, stat.IsDir())
-
-	stat, err = os.Stat(filepath.Join(dstDir, childDir, childTempFile))
-	assert.NoError(t, err)
-	assert.Equal(t, uint32(0755), uint32(stat.Mode()))
-	assert.False(t, stat.IsDir())
-	data, err = os.ReadFile(filepath.Join(dstDir, childDir, childTempFile))
-	assert.NoError(t, err)
-	assert.Equal(t, "test file", string(data))
-
-	err = CopyDir(srcDir, dstDir)
-	assert.Error(t, err)
-}
 func TestLimitedReaderWithError(t *testing.T) {
 	t.Run("read less than max size", func(t *testing.T) {
 		maxBytes := 10
